+++ front/src/pages/Settings/BasicSettings/styles.ts
